src/requirements.txt
aiodns
aiohttp
asyncpg
dateparser
discord.py
humanize
uvloop
ciso8601
pytz
python-dotenv